% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knvb_cols}
\alias{knvb_cols}
\title{Dataframe with KNVB colors used in palette functions}
\format{
\subsection{\code{knvb_cols}}{

#' A data frame with 18 rows and 2 columns:
\describe{
\item{name}{Color name}
\item{value}{RGB hex value}
}
}
}
\usage{
knvb_cols
}
\description{
Dataframe with KNVB colors used in palette functions
}
\keyword{datasets}
