% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knvb_cols_ls}
\alias{knvb_cols_ls}
\title{List with KNVB colors, used to set colors manually}
\format{
An object of class \code{list} of length 18.
}
\usage{
knvb_cols_ls
}
\description{
List with KNVB colors, used to set colors manually
}
\keyword{datasets}
