% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knvb.R
\name{theme_knvb}
\alias{theme_knvb}
\title{KNVB theme}
\usage{
theme_knvb(
           base_size = 12,
           y_grid = T,
           x_grid = T,
           ...
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{y_grid}{logical. Indicate whether you want to draw major and minor
y-axis grid lines}

\item{x_grid}{logical. Indicate whether you want to draw major and minor
x-axis grid lines}

\item{...}{potential further arguments passed to \code{\link[=theme_grey]{theme_grey()}}}
}
\description{
Custom \link{ggplot2} theme used in all KNVB figures. Created by KNVB
research department.
}
\examples{
library(ggplot2)

mtcars2 <- within(mtcars, {
  vs <- factor(vs, labels = c("V-shaped", "Straight"))
  am <- factor(am, labels = c("Automatic", "Manual"))
  cyl  <- factor(cyl)
  gear <- factor(gear)
})

p1 <- ggplot(mtcars2) +
  geom_point(aes(x = wt, y = mpg, colour = gear)) +
  labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "(1973-74)",
    caption = "Data from the 1974 Motor Trend US magazine.",
    x = "Weight (1000 lbs)",
    y = "Fuel economy (mpg)",
    colour = "Gears"
  )

p1 + theme_knvb()

}
