% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knvb.R
\name{theme_knvb}
\alias{theme_knvb}
\title{KNVB ggplot2 theme}
\usage{
theme_knvb(base_size = 12, base_family = "Arial")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Custom ggplot2 theme used in all KNVB figures created by KNVB research department
}
